Extract pipx_installed to standalone
