+++ plextraktsync/commands/self_update.py
@@ -1,36 +1,11 @@
-import json
-from json import JSONDecodeError
 from os import system
 
 import click
 
-from plextraktsync.util.execx import execx
-
-
-def pipx_installed(package: str):
-    try:
-        output = execx("pipx list --json")
-    except FileNotFoundError:
-        return None
-    if not output:
-        return None
-
-    try:
-        install_data = json.loads(output)
-    except JSONDecodeError:
-        return None
-    if install_data is None:
-        return None
-
-    try:
-        package = install_data["venvs"][package]["metadata"]["main_package"]
-    except KeyError:
-        return None
-
-    return package
-
 
 def enable_self_update():
+    from plextraktsync.util.packaging import pipx_installed
+
     package = pipx_installed("plextraktsync")
 
     return package is not None
