+++ plextraktsync/util/packaging.py
@@ -1,5 +1,9 @@
+import json
 import site
+from json import JSONDecodeError
 from os.path import dirname
+
+from plextraktsync.util.execx import execx
 
 
 def installed():
@@ -10,3 +14,26 @@ def installed():
     paths = site.getsitepackages()
 
     return absdir in paths
+
+
+def pipx_installed(package: str):
+    try:
+        output = execx("pipx list --json")
+    except FileNotFoundError:
+        return None
+    if not output:
+        return None
+
+    try:
+        install_data = json.loads(output)
+    except JSONDecodeError:
+        return None
+    if install_data is None:
+        return None
+
+    try:
+        package = install_data["venvs"][package]["metadata"]["main_package"]
+    except KeyError:
+        return None
+
+    return package
